Move observer and database into cli_extras
